+++ src/pages/Board.jsx
@@ -19,15 +19,6 @@ const Board = () => {
   const [board, setBoard] = React.useState({})
   const toast = useToast()
 
-  if (!boardID.length) {
-    toast({
-      status: 'error',
-      title: 'No boards',
-      description: "Let's create a new board!",
-      duration: 3000,
-    })
-  }
-
   const fetchData = React.useCallback(async () => {
     const data = await getBoard(boardID)
     setBoard(data)
@@ -44,59 +35,83 @@ const Board = () => {
   return (
     <Box h="100vh">
       <Center>
-        <Box>
+        <Box mt="10">
           <SimpleForm
             buttonText="Add task group"
             inputPlaceholder="Name of the new task group..."
             onFormSubmit={async (value) => {
               await createTaskGroup(Number(boardID), value)
               fetchData()
+              toast({
+                status: 'success',
+                title: 'Successfully',
+                description: `Added task group: (${value.toUpperCase()})`,
+                duration: 3000,
+              })
             }}
             onSubmit={async (newBoardName) => {
               await updateBoard(Number(boardID), { name: newBoardName })
               fetchData()
             }}
           />
-          <Text color="gray.400" mt="5">
-            {boardID.length} groups
-          </Text>
           <Center>
             <SimpleGrid mt="10" columns="2" spacingX="20px" spacingY="20px">
               {
                 // eslint-disable-next-line no-shadow
                 board.taskGroups?.map(({ name: newName, id: taskGroupID, taskIds }) => (
-                  <TaskGroup
-                    key={taskGroupID}
-                    title={newName}
-                    color="blue.500"
-                    onDelete={async () => {
-                      await removeTaskGroup(taskGroupID)
-                      fetchData()
-                    }}
-                    onSubmit={async (newTaskGroupName) => {
-                      await updateTaskGroup(taskGroupID, { name: newTaskGroupName })
-                      fetchData()
-                    }}
-                  >
-                    {board.tasks
-                      ?.filter((tasks) => taskIds.includes(tasks.id))
-                      // eslint-disable-next-line no-shadow
-                      .map(({ id: taskID, name, content }) => (
-                        <Task key={taskID} name={name} content={content} />
-                      ))}
-                    <SimpleForm
-                      buttonText="Add"
-                      inputPlaceholder="Task name"
-                      onFormSubmit={async (value) => {
-                        await createTask(Number(boardID), Number(taskGroupID), { name: value })
+                  <>
+                    <Text fontWeight="thin" color="gray.400" position="absolute" mt="-35">
+                      {board.taskGroups.length} groups
+                    </Text>
+                    <TaskGroup
+                      key={taskGroupID}
+                      title={newName}
+                      color="blue.500"
+                      onDelete={async () => {
+                        await removeTaskGroup(taskGroupID)
+                        fetchData()
+                        toast({
+                          status: 'error',
+                          title: `Delete group: ${newName.toUpperCase()}`,
+                          description: "Let's create a new group!",
+                          duration: 3000,
+                        })
+                      }}
+                      onSubmit={async (newTaskGroupName) => {
+                        await updateTaskGroup(taskGroupID, { name: newTaskGroupName })
                         fetchData()
                       }}
-                      onSubmit={async (newTaskName) => {
-                        await updateTask({ name: newTaskName })
-                        fetchData()
-                      }}
-                    />
-                  </TaskGroup>
+                    >
+                      {board.tasks
+                        ?.filter((tasks) => taskIds.includes(tasks.id))
+                        // eslint-disable-next-line no-shadow
+                        .map(({ id: taskID, name, content }) => (
+                          <Task key={taskID} name={name} content={content} />
+                        ))}
+
+                      <Text color="gray.400" mt="5">
+                        {taskIds.length} tasks
+                      </Text>
+                      <SimpleForm
+                        buttonText="Add"
+                        inputPlaceholder="Task name"
+                        onFormSubmit={async (value) => {
+                          await createTask(Number(boardID), Number(taskGroupID), { name: value })
+                          fetchData()
+                          toast({
+                            status: 'success',
+                            title: 'Successfully',
+                            description: `Added task: (${value.toUpperCase()})`,
+                            duration: 3000,
+                          })
+                        }}
+                        onSubmit={async (newTaskName) => {
+                          await updateTask(taskIds, { name: newTaskName })
+                          fetchData()
+                        }}
+                      />
+                    </TaskGroup>
+                  </>
                 ))
               }
             </SimpleGrid>
